+++ apps/api/infra/api.ts
@@ -150,6 +150,7 @@ export const serviceSetup = (services: {
       RSK_API_PASSWORD: '/k8s/shared/api/RSK_API_PASSWORD',
       RSK_API_URL: '/k8s/shared/api/RSK_API_URL',
       ISLYKILL_SERVICE_PASSPHRASE: '/k8s/api/ISLYKILL_SERVICE_PASSPHRASE',
+      XROAD_NATIONAL_REGISTRY_TIMEOUT: '20000',
     })
     .xroad(
       Base,
