+++ infra/src/dsl/xroad.ts
@@ -1,9 +1,4 @@
-import {
-  EnvironmentVariables,
-  MissingSetting,
-  Secrets,
-  XroadConfig,
-} from './types/input-types'
+import { EnvironmentVariables, Secrets, XroadConfig } from './types/input-types'
 
 type XroadSectionConfig = {
   secrets?: Secrets
