+++ libs/api/domains/national-registry-x-road/src/models/nationalRegistryPerson.model.ts
@@ -15,7 +15,7 @@ export class NationalRegistryPerson {
   genderCode!: string
 
   @Field(() => NationalRegistryAddress, { nullable: true })
-  address?: NationalRegistryAddress
+  address?: NationalRegistryAddress | null
 
   @Field(() => Boolean, { nullable: true })
   livesWithApplicant?: boolean
