+++ libs/clients/middlewares/src/index.ts
@@ -2,5 +2,5 @@ export * from './lib/createEnhancedFetch'
 export * from './lib/types'
 export { Request, Response } from './lib/nodeFetch'
 export { FetchError } from './lib/FetchError'
-export * from './lib/buildCacheControl'
-export { CacheConfig } from './lib/withCache'
+export * from './lib/withCache/buildCacheControl'
+export { CacheConfig } from './lib/withCache/types'
